backend/requirements.txt
uvicorn[standard]
fastapi
mcstatus
apscheduler
pytz
python-dateutil
aiohttp